Move donation bloc into its own folder
